+++ HISTORY.rst
@@ -35,6 +35,7 @@
 - When changing document types, don't delete the old metadata that is
   also found in the new document type. GitLab issue #421.
 - Add tag attach and tag remove events.
+- Change the permission needed to attach and remove tags.
 
 2.6.4 (2017-07-26)
 ==================